Support anchors that are not the `parent` node.
There can be situations where the element the FlowTip™ is anchored to is not the bounding parent. This happens in any situation where there is a `position: "relative"` element along the hierarchy up to the `parent`. Fixing this is slightly less than trivial, but the solution offers some nice perks.

Instead of using offsets relative to the parent, always use a single coordinate system via `getBoundingClientRect()`. This returns _any_ elements position relative to the viewport, not to its parent – we have a unique single frame of reference from which all other calculations can be made. If the parent is the same as the anchor, no changes need to be made when doing the calculations. However, if the parent is not the anchor, then we simply "denormalize" the resulting calculations by shifting the bounding box the distance between the parent and the anchor. This has the net effect of translating the FlowTip™ element by the amount needed to essentially make it look like the anchor had no role in its positioning in the first place.

As far as the implementation goes, it requires that edge checks are no longer 0-based since the edge of the bounding box can be negative. For these cases adjustments from the parent bounding box have been added.

The demo has been updated to showcase these new features.