Stateless refactor.
Instead of having FlowTip™ manage DOM details (like getting the width/height of an element) make FlowTip™ simply a collection of mathematical utilities that, given some rects, calculates where the tooltip and tail should be displayed. It is then up to the child function of the FlowTip™ element to generate the correct elements for display. This has the net effect of reducing the complexity of FlowTip™ and allowing support for platforms other than just the DOM. One downside of this is that the region is no longer sticky – effectively the old `perservere` option is now always true and its up to someone else to manage keeping the region in sync.

The DOM specific code has been moved into `dom.js`. This higher-order component takes component classes for `content`, `tail` and `anchor` and returns a new class that handles all the DOM rangling for a FlowTip™ that uses the aforementioned classes. Unfortunately due the way the resize watching mechanism works both `tail` and `content` must be component classes right now and not stateless.

The overall demo code is significantly smaller and cleaner. The `webpack` configuration has been updated to build the `dom.js` bits. Eventually these bits should probably be moved to a `flowtip-dom` akin to `react-dom`.

Additionally, provided the opportunity for such large changes, use the MetaLab `eslint` configuration and use `babel` for building the generated files instead of `webpack`.

Tails are also now bound to the edge of the content container.