Use `window` dimensions instead of `screen` ones.
When trying to keep the flyout in the viewport, `screen` does not accurately represent the viewport – it represents the physical screen size. This is doubly true on mobile devices that switch to portrait mode where the `width` and `height` do not change with respect to the orientation change. Because of this, using the `window.inner*` properties make a lot more sense.

This closes #19.