Fix incorrect edge detection logic.
It looks like previously `.indexOf !== region` should have been `.indexOf(region) !== -1`. Since incorrect edge flipping to the point of UI breakage doesn't happen that often this went undetected for quite some time. So the logic is now fixed to do the right thing and does so in a way a little more performant than calling `indexOf` on a two-element constant array.