Improve support for parent container detection
A mounted instance of `FlowTip` has two important ancestor elements that define how it is positioned. It has a "containing block" and a "clipping block", which most of the time are the `body` element.

In circumstances where `flowTip` has an ancestor element with an `overflow` other than `visible`, this is detected an set as the clipping block. Its dimensions are used as the `bounds` rect during positioning. This ensures that a `FlowTip` instance will be constrained relative to any element that represents a visibility boundary.

The "containing block" is the closest ancestor element with `position` other than static. The `style` returned to the `content` component contains a `top` `left` position relative to this element.