Include edgeOffset in layout calculation
`edgeOffset` is used to represent a margin at the the bounds rect edge. Previously it was only a prop used within the `FlowTip` react component that was subtracted from the resolved `bounds` rect, effectively shrinking it. This had the undesirable effect of also applying the margin to between the positioned tooltip and viewport edge when the target was outside the viewport.

What we really want the tail side of the tooltip to have its own margin, separate from the body. When the tooltip is positioned at the screen edge, the larger of the two margins will determine the final position.