Fix typescript migration issues
Configure typescript to be as strict as possible and fix the issues that this surfaces.

Use the tsconfig extends feature to move as all shared typescript config to the repo root.

Add the missing `types` fields to each `package.json` so the types will actually be available to package consumers.

Move all build results to `lib` directories for clarity and convenience.